merges test/annotations.ts into test/IonAnnotations.ts
